requirements.txt
Flask
python-dotenv
google-generativeai
plotly
requests
lxml
markdown-it-py
werkzeug